refactor(media): split transcription persistence support
